@@ -18,7 +18,7 @@ const (
 	configFileName = "gitops-promotion.yaml"
 )
 
-// nolint:funlen,cyclop,gocognit // ignore
+//nolint:funlen,cyclop,gocognit // ignore
 func Run(ctx context.Context, args []string) (string, error) {
 	if len(args) < 2 {
 		return "", fmt.Errorf("new, feature, promote, or status subcommand is required")
