@@ -167,7 +167,8 @@ func (g *Repository) CreateCommit(branchName, message string) (*git2go.Oid, erro
 
 // GetLastCommitForPath returns the last commit for the given path. All files and subdirectories
 // will be considered if the path is a directory.
-// nolint:gocognit // ignore
+//
+//nolint:gocognit // ignore
 func (g *Repository) GetLastCommitForPath(path string) (*git2go.Commit, error) {
 	// There is currently no implementation in libgit2 to recursively check files.
 	// Here is an issue in git2go which discusses the problem of recursively checking files.
